requirements.txt
fastapi
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
python-json-logger
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp